Use solid-client-authn-browser if possible
This inspects the current dependencies, and if
solid-client-authn-browser is available, it uses the default session to
get an authenticated fetch.

This is helpful because the VC issuer will likely expect calls to its
API to be authenticated, which would de facto make the `fetch`
parameter mandatory, and this allows to make it optional.