Work around rollup ts plugin ES issue
This prevents the bahaviour documented in
https://github.com/ezolenko/rollup-plugin-typescript2/pull/170 from
happening. Namely, a _virtual/tslib.js file would be emitted, which
confuses NodeJS because we expect all ES modules to have the .mjs
extension.

The previous syntax feels more elegant though, so I think it should be
re-enabled once the bug is resolved. Also, there's no guarantee that
this issue can alway be worked around so easily. Issue tracking is
available at
https://github.com/ezolenko/rollup-plugin-typescript2/issues/282.