Explicitly fetch config discovery as JSON-LD
The configuration discovery document (.well-known/vc-configuration) is only available as JSON-LD, and trying to fetch it with the default behavior of `getSolidDataset` fails with '406 unacceptable', because it tries to dereference the document as Turtle. This upgrades to the latest version of `@inrupt/solid-client`, which exports functions to explicitly parse a dataset as JSON-LD, and uses said function when looking up the configuration discovery document.