Add includeExpiredVc option to VC lookup (#303)
`getVerifiableCredentialAllFromShape` now supports a new option, `includeExpiredVc`. If set to true, the Holder endpoint will return VCs that have expired and are thus no longer valid. This new option defaults to false. Note that this behaviour is off-standard, and relies on our own implementation. Striclty specification-conforming VC holders will not consider this option (but they will not break on it either).