Accept two possible outcomes of issuer bad input
When being sent a VC which doesn't match any of the preconfigured shapes, the issuer may have one of two behavours: it may either return an error (400 Bad Request), or return a VC (200 OK) without any "useful" type, which are attached to the shapes. This adjusts the tests so that both are supported by the end-to-end test. In the future, it should be driven by configuration what the current behaviour is expected to be as part of the e2e test setup.