chore: Omnibus PR (425, 418, 426)
Combines the following PRs:

- https://github.com/inrupt/solid-client-vc-js/pull/425
- https://github.com/inrupt/solid-client-vc-js/pull/418
- https://github.com/inrupt/solid-client-vc-js/pull/426

Also adds:

- GitHub Actions checks for running our standard audit and lint actions.
- Fixed remaining issues in adopting our style guides (headers, prettier)
- Fixed linting of node e2e tests

Commits:

* chore: upgrade typescript and shared eslint configuration

* chore: upgrade inrupt SDKs

* chore: improve strictness of the engines supported

This explicitly disables Node 18, as we don't yet support it

* ci: disable browser-based e2e tests as they do nothing and use testcafe

* chore: remove stubbed out testcafe e2e browser tests

Browser E2E tests were never implemented in this library, nor was it migrated to Playwright, this clears the way for setting that up.

* chore: disable node e2e tests that point to ESS 1.1

* tests: Accept two possible outcomes of issuer bad input

When being sent a VC which doesn't match any of the preconfigured shapes, the issuer may have one of two behavours: it may either return an error (400 Bad Request), or return a VC (200 OK) without any "useful" type, which are attached to the shapes. This adjusts the tests so that both are supported by the end-to-end test. In the future, it should be driven by configuration what the current behaviour is expected to be as part of the e2e test setup.

* chore: upgrade typedoc to ^0.23.8

* ci: add shared lint and audit actions/jobs

* chore: use more standard linting setup

* chore: run lint:fix for header and prettier

Co-authored-by: Nicolas Seydoux <nseydoux@inrupt.com>