Add support for Verifiable Presentations Requests
The legacy VC lookup mechanism relies on an ad-hoc object, and uses the /derive endpoint, which is actually not meant for querying but rather for selective disclosure. This introduces a query function which implements the Verifiable Presentation Request mechanism as described in https://w3c-ccg.github.io/vp-request-spec/. An important note is that we make the assumption that an endpoint supporting Verifiable Presentation Request is available at a /query path, which is outside of the VC API specification scope. This assumption is used to distinguish new endpoints vs legacy endpoints.

Currently, only Query by Example VPRs are supported, which is similar to the legacy behavior in a lot of ways. The legacy getVerifiableCredentialAllFromShape function now supports both legacy and VPR-compliant endpoints.